Update for 5 dice

+++ A5/A5/Camera.h
@@ -126,7 +126,7 @@ private:
 	GLfloat radians(GLfloat degrees)
 	{
 		// TODO: Make sure this works correctly
-		return (degrees * ((2.0 * PI) / 360.0));
+		return (degrees * ((2.0f * (float)PI) / 360.0f));
 	}
 	
 };
