+++ A5/A5/PlaneRotation.h
@@ -13,7 +13,7 @@
 
 GLfloat radians(GLfloat degrees)
 {
-	return (degrees * ((2.0 * PI) / 360.0));
+	return (degrees * ((2.0f * (float)PI) / 360.0f));
 }
 
 mat4 getRotationMatrix(GLfloat yawR, GLfloat pitchR, GLfloat rollR, vec4 &fV, vec4 &rightV, vec4 &upV)
